backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlmodel>=0.0.14
pydantic[email]>=2.4.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
plaid-python>=11.0.0
python-dotenv>=1.0.0
pytest>=7.4.0
httpx>=0.25.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
cryptography>=41.0.0
requests>=2.31.0